# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/v1/wizard":{"get":{"operationId":"WizardController_findAll","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Wizards retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Paginate"}}}}},"tags":["Wizards"]},"post":{"operationId":"WizardController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWizardDto"}}}},"responses":{"201":{"description":"Wizard created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Wizards"]}},"/api/v1/wizard/{id}":{"get":{"operationId":"WizardController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Wizard retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Wizards"]},"patch":{"operationId":"WizardController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWizardDto"}}}},"responses":{"200":{"description":"Wizard Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Wizards"]},"delete":{"operationId":"WizardController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Wizard deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Wizards"]}},"/api/v1/wizard/assign":{"patch":{"operationId":"WizardController_assign","parameters":[],"requestBody":{"required":true,"description":"Assign a spell to a wizard","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignSpellToWizardDto"}}}},"responses":{"200":{"description":"Spell Assigned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Wizards"]}},"/api/v1/spell":{"get":{"operationId":"SpellController_findAll","parameters":[{"name":"name","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Spells retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpellDto"}}}}}},"tags":["Spells"]},"post":{"operationId":"SpellController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSpellDto"}}}},"responses":{"201":{"description":"Spell created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Spells"]}},"/api/v1/spell/{id}":{"get":{"operationId":"SpellController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Spell retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Spells"]},"patch":{"operationId":"SpellController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSpellDto"}}}},"responses":{"200":{"description":"Spell Updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Spells"]},"delete":{"operationId":"SpellController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Spell deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandardResponse"}}}}},"tags":["Spells"]}},"/api/v1/elixirs":{"post":{"operationId":"ElixirsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateElixirDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ElixirsController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/elixirs/{id}":{"get":{"operationId":"ElixirsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ElixirsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateElixirDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ElixirsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Wizard World API","description":"Huguoh Backend Developer Challenge","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8989"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"meta":{"type":"object","properties":{"count":{"type":"number"},"currentPage":{"type":"number"},"nextPage":{"type":"number"},"prevPage":{"type":"number"},"lastPage":{"type":"number"}},"required":["count","currentPage","nextPage","prevPage","lastPage"]},"Paginate":{"type":"object","properties":{"status":{"type":"string"},"data":{"type":"array","items":{"type":"string"}},"meta":{"$ref":"#/components/schemas/meta"}},"required":["status","data","meta"]},"CreateWizardDto":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"}},"required":["firstname","lastname"]},"StandardResponse":{"type":"object","properties":{"status":{"type":"string"},"code":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"},"meta":{"type":"object"}},"required":["status","code","message","data","meta"]},"UpdateWizardDto":{"type":"object","properties":{}},"AssignSpellToWizardDto":{"type":"object","properties":{"wizard_id":{"type":"string","example":"40d61093-2fcb-424c-be4c-413dcfc519fb"},"spell_id":{"type":"string","example":"40d61093-2fcb-424c-be4c-413dcfc519fb"}},"required":["wizard_id","spell_id"]},"SpellDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"incantation":{"type":"string"},"effect":{"type":"string"},"creator":{"type":"string"},"canBeVerbal":{"type":"boolean"},"type":{"type":"string","enum":["None","Charm","Conjuration","Spell","Transfiguration","Healing Spell","Dark Charm","Jinx","Curse","Magical Transportation","Hex","Counter Spell","Dark Arts","Counter Jinx","Counter Charm","Untransfiguration","Binding MagicalContract","Vanishment"]},"light":{"type":"string","enum":["None","Blue","IcyBlue","Red","Gold","Purple","Transparent","White","Green","Orange","Yellow","BrightBlue","Pink","Violet","BlueishWhite","Silver","Scarlet","Fire","FieryScarlet","Grey","DarkRed","Turquoise","PsychedelicTransparentWave","BrightYellow","BlackSmoke"]}},"required":["id","name","incantation","effect","creator","canBeVerbal","type","light"]},"CreateSpellDto":{"type":"object","properties":{"name":{"type":"string"},"incantation":{"type":"string"},"effect":{"type":"string"},"creator":{"type":"string"},"canBeVerbal":{"type":"boolean"},"type":{"type":"string","enum":["None","Charm","Conjuration","Spell","Transfiguration","Healing Spell","Dark Charm","Jinx","Curse","Magical Transportation","Hex","Counter Spell","Dark Arts","Counter Jinx","Counter Charm","Untransfiguration","Binding MagicalContract","Vanishment"]},"light":{"type":"string","enum":["None","Blue","IcyBlue","Red","Gold","Purple","Transparent","White","Green","Orange","Yellow","BrightBlue","Pink","Violet","BlueishWhite","Silver","Scarlet","Fire","FieryScarlet","Grey","DarkRed","Turquoise","PsychedelicTransparentWave","BrightYellow","BlackSmoke"]}},"required":["name","incantation","effect","creator","canBeVerbal","type","light"]},"UpdateSpellDto":{"type":"object","properties":{"name":{"type":"string"},"incantation":{"type":"string"},"effect":{"type":"string"},"creator":{"type":"string"},"canBeVerbal":{"type":"boolean"},"type":{"type":"string","enum":["None","Charm","Conjuration","Spell","Transfiguration","Healing Spell","Dark Charm","Jinx","Curse","Magical Transportation","Hex","Counter Spell","Dark Arts","Counter Jinx","Counter Charm","Untransfiguration","Binding MagicalContract","Vanishment"]},"light":{"type":"string","enum":["None","Blue","IcyBlue","Red","Gold","Purple","Transparent","White","Green","Orange","Yellow","BrightBlue","Pink","Violet","BlueishWhite","Silver","Scarlet","Fire","FieryScarlet","Grey","DarkRed","Turquoise","PsychedelicTransparentWave","BrightYellow","BlackSmoke"]}}},"CreateElixirDto":{"type":"object","properties":{}},"UpdateElixirDto":{"type":"object","properties":{}}}}}